+++ R/tm_t_rsp.R
@@ -100,7 +100,6 @@ template_rsp <- function(dataname,
     data_list,
     substitute_names(
       expr = dplyr::mutate(is_rsp = aval_var %in% responder_val) %>%
-        dplyr::filter(aval %in% responder_val) %>%
         dplyr::mutate(aval = factor(aval_var, levels = responder_val)),
       names = list(
         aval = as.name(aval_var)
@@ -111,8 +110,6 @@ template_rsp <- function(dataname,
       )
     )
   )
-
-  data_list <- add_expr(data_list, quote(df_explicit_na()))
 
   y$data <- substitute(
     expr = {
@@ -437,7 +434,7 @@ template_rsp <- function(dataname,
 #'         BESRSPI = list(
 #'           rsp = c("Complete Response (CR)", "Partial Response (PR)"),
 #'           levels = c("Complete Response (CR)", "Partial Response (PR)",
-#'              "Stable Disease (SD)", "Partial Disease (PD)")),
+#'              "Stable Disease (SD)", "Progressive Disease (PD)")),
 #'         INVET = list(rsp = c("Stable Disease (SD)"),
 #'           levels = c("Complete Response (CR)", "Partial Response (PR)", "Stable Disease (SD)")),
 #'         OVRINF = list(rsp = c("Partial Response (PR)")))
@@ -462,7 +459,7 @@ tm_t_rsp <- function(label,
                      strata_var,
                      aval_var = choices_selected(variable_choices(dataname, "AVALC"), "AVALC", fixed = TRUE),
                      conf_level = choices_selected(c(0.95, 0.9, 0.8), 0.95, keep_order = TRUE),
-                     default_responses = NULL,
+                     default_responses = c("CR", "PR", "Y", "Complete Response (CR)", "Partial Response (PR)"),
                      add_total = FALSE,
                      pre_output = NULL,
                      post_output = NULL) {
@@ -678,25 +675,15 @@ srv_t_rsp <- function(input,
         sel_param <- if (is.list(default_responses)) {
           default_responses[[input[[extract_input("paramcd", paramcd$filter[[1]]$dataname, filter = TRUE)]]]]
         } else default_responses
-        common_rsp <- if (is.list(default_responses)) {
-          if (is.list(sel_param)) {
-            sel_param$rsp
-          } else {
-            sel_param
-          }
-        } else {
-          c("CR", "PR", "Y", "Complete Response (CR)", "Partial Response (PR)")
-        }
+        common_rsp <- if (is.list(sel_param)) {
+          sel_param$rsp
+        } else sel_param
         responder_choices <- if (is_empty(aval_var)) {
           character(0)
         } else {
           if ("levels" %in% names(sel_param)) {
-            if (length(intersect(unique(anl_merged()$data()[[aval_var]]), sel_param$levels)) > 1) {
-              sel_param$levels
-            }
-          } else {
-            unique(anl_merged()$data()[[aval_var]])
-          }
+            sel_param$levels
+          } else unique(anl_merged()$data()[[aval_var]])
         }
         updateSelectInput(
           session, "responders",
@@ -767,25 +754,30 @@ srv_t_rsp <- function(input,
       }
     )
 
-    validate(
-      need(all(unlist(lapply(default_responses, function(x) {
-        browser()
-        if (is.list(x)) {
-          if (length(x) == 2) {
-            all(names(x) %in% c("rsp", "levels"))
-          } else TRUE
-        } else TRUE}))),
-        "The lists given in default_responses must contain named lists 'rsp' and 'levels'.")
-    )
+    if (!identical(default_responses, c("CR", "PR", "Y", "Complete Response (CR)", "Partial Response (PR)"))) {
+      validate(
+        need(all(unlist(lapply(default_responses, function(x) {
+          if (is.list(x) & "levels" %in% names(x)) {
+            lvls <- x$levels
+            all(x$rsp %in% lvls)
+          } else {
+            lvls <- unique(anl_merged()$data()[[input$`aval_var-dataset_ADRS_singleextract-select`]])
+            if ("rsp" %in% names(x)) {
+              all(x$rsp %in% lvls)
+            } else all(x %in% lvls)
+          }
+          }))),
+          "All selected default responses must be in the levels of AVAL.")
+      )
+    }
 
-    validate(
-      need(all(unlist(lapply(default_responses, function(x) {
-        lvls <- if (is.list(x)) x$levels else NULL
-        if (is.null(lvls)) {
-          all(unlist(x) %in% unique(unlist(anl_merged()$data()[c(aval_var$select$choices)])))
-        } else TRUE}))),
-        "All selected default responses must be in AVAL")
-    )
+    if (is.list(default_responses)) {
+      validate(
+        need(all(grepl("\\.rsp|\\.levels", names(unlist(default_responses))) |
+                   names(unlist(default_responses)) %in% names(default_responses)),
+             "The lists given for each AVAL in default_responses must be named 'rsp' and 'levels'.")
+      )
+    }
 
     validate(need(
       input$conf_level >= 0 && input$conf_level <= 1,